moved all iterators for rebuilding index to a separate file
